Refactor training into ACXTrainer class
